refactor: extract query and match expression compilers, consolidate helpers, and update related benchmarks
